requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.9
aiofiles==24.1.0
requests==2.32.3

# RAG & Embeddings
ollama==0.3.3
qdrant-client==1.12.1
langchain-text-splitters==0.3.2

# Document Parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Config & Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
